# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Tasks Demo Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/tasks": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Get List of Tasks",
                "operationId": "feed89770241fcb6af2325dafc9adfde",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status",
                        "required": false,
                        "example": 0
                    },
                    {
                        "name": "due_date",
                        "in": "query",
                        "description": "Due date",
                        "required": false,
                        "example": "year"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "New Title"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Some Description here"
                                                    },
                                                    "status": {
                                                        "type": "integer",
                                                        "example": 0
                                                    },
                                                    "due_date": {
                                                        "type": "string",
                                                        "example": "2024-08-04 13:33:03"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2024-08-01 13:33:03"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Store New Task",
                "operationId": "44140ecf3dfc93dfa6ae24da60ceb875",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "integer"
                                    },
                                    "due_date": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "New Title",
                                    "description": "Some Description here",
                                    "status": 0,
                                    "due_date": "2024-08-04 13:33:03"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "New Title"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Some Description here"
                                                },
                                                "status": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "due_date": {
                                                    "type": "string",
                                                    "example": "2024-08-04 13:33:03"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-08-01 13:33:03"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tasks/{task}": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Get Task Information",
                "operationId": "9ccc9fcdee5d706b5aad72d0502c2092",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "Task id",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "New Title"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Some Description here"
                                                },
                                                "status": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "due_date": {
                                                    "type": "string",
                                                    "example": "2024-08-04 13:33:03"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-08-01 13:33:03"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Task",
            "description": "API Endpoints of Tasks"
        }
    ]
}